Add [all] dependency to clarify napari and Qt binding installation (#56)
Closes #52. Partly addresses https://github.com/napari/docs/issues/131.

Provides a single [all] optional dependency to install napari[all] --
bringing in both napari and the default Qt binding. This accomplishes a
few things:

1. Signals to plugin builders that Qt bindings (and to a lesser extent,
napari) need to be an _optional_ dependency. Without this implied
install, I and many others stick napari[all] in the main dependencies.
2. Provides a one-line install for users `pip install plugin_name[all]`
to get the plugin + napari + qt

~~In addition, this also adds a _qt_binding_check file modeled after
napari's to give more verbose output when Qt bindings are absent. It
will only be added if a widget is added.~~

---------

Co-authored-by: Carol Willing <carolcode@willingconsulting.com>