Remove default categories from napari.yaml (#72)
By default, three categories are given to `napari.yaml` that might not
be appropriate for the plugin. This PR removes those categories, but
keeps this information as a comment to prompt plugin creators to add
categories appropriate to their plugin. This prevents new plugins be
categorized the same if an author does not update this field.