Genereate .copier-answer.yml and add template tag to README (#78)
# Description

This genarates the `.copier-answer.yml` file which I've been debating on
adding for a while. My hesitation is adding yet another file into the
template's output, but I'm realizing a few strengths:
1. authors can validate how they set up a project, and use that in a new
project, if they find it helpful
2. support, like from the napari team, will be easier because we can see
what the project started with, in order to narrow down changes
3. it allows re-running templates and updating, by basing on previous
answers or 'resetting'

This also adds the `vcs_ref` (tag or dev version) to the template
output, so that, again, it's easier to support plugin authors.