switching from `project.optional-dependencies` to `dependency-groups` (#93)
# References and relevant issues

closes https://github.com/napari/napari-plugin-template/issues/92

ref. https://peps.python.org/pep-0735/

<!-- What relevant resources were used in the creation of this PR?
If this PR addresses an existing issue on the repo,
please link to that issue here as "Closes #(issue-number)"
If this PR depends on another PR/issue (even in another repo),
please link to it with "Depends on #PR-number" or "Depends on
org/repo#PR-number"
-->

# Description

switching from `project.optional-dependencies` to `dependency-groups`

<!-- What does this pull request (PR) do? Is it a new feature, a bug
fix,
an improvement, or something else? Why is it necessary? If relevant,
please add
a screenshot or a screen capture: "An image is worth a thousand words!"
-->

---------

Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>