Improve napari_get_reader to check dtype without loading full array (#108)
# References and relevant issues
Closes #103

# Description
This PR removes the file extension check in the `napari_get_reader`
function, because this is now handled by `filename_patterns`.

It also adds another check for the actual format of the file, without
loading the file into memory, and adds a comment about the purpose of
`napari_get_reader`.

---------

Co-authored-by: Draga Doncila <ddon0001@student.monash.edu>